+++ Asset Management Platform/SecurityClasses/FixedIncome.cs
@@ -22,13 +22,13 @@ namespace Asset_Management_Platform
         }
 
 
-        public FixedIncome(string cusip, string ticker, string description, double lastPrice, double yield) 
+        public FixedIncome(string cusip, string ticker, string description, float lastPrice, double yield) 
             : base(cusip, ticker, description, lastPrice, yield)
         {
 
         }
 
-        public FixedIncome(string cusip, string ticker, string description, double lastPrice, double yield, double coupon, string issuer, string rating)
+        public FixedIncome(string cusip, string ticker, string description, float lastPrice, double yield, double coupon, string issuer, string rating)
             : base(cusip, ticker, description, lastPrice, yield)
         {
             _coupon = coupon;
