+++ Asset Management Platform/SecurityClasses/MunicipalBond.cs
@@ -8,13 +8,13 @@ namespace Asset_Management_Platform.SecurityClasses
 {
     class MunicipalBond : FixedIncome
     {
-        public MunicipalBond(string cusip, string ticker, string description, double lastPrice, double yield) 
+        public MunicipalBond(string cusip, string ticker, string description, float lastPrice, double yield) 
             : base(cusip, ticker, description, lastPrice, yield)
         {
             SecurityType = "Municipal Bond";
         }
 
-        public MunicipalBond(string cusip, string ticker, string description, double lastPrice, double yield, double coupon, string issuer, string rating)
+        public MunicipalBond(string cusip, string ticker, string description, float lastPrice, double yield, double coupon, string issuer, string rating)
             : base(cusip, ticker, description, lastPrice, yield, coupon, issuer, rating)
         {
             SecurityType = "Municipal Bond";
